Move rdb to /rdb/, move crc64 to /rdb/crc64/, move main to /cmd/redis-resharding-proxy/
