+++ img/redis-resharding.png
